be/requirements.txt
flask
flask-cors
pymysql
requests
python-dotenv
rapidfuzz
scikit-learn
pandas
